Add working solver for the Grecian Computer puzzle
The `Puzzle` class takes a set of `DiscStack`s and sets the `shifts=`
value, being all the permutations (with repitition) of the integers 0..11,
of length 5. The puzzle is solved when all the columns sum to 42.

There's a `DiscStack#rotate!` method to rotate a single disc in the
stack. Note that currently, if you rotate a disc that way and then try
to set the shifts with the Puzzle it probably won't work. I think you'll
need ot reset the DiscStack back to the base state (where shifts = all
zeros).